Removed not-really-needed Flask-Restful dependency
